requirements.txt
Pillow==10.1.0
pymediainfo==6.1.0
